+++ packages/aws-amplify-react/src/Auth/index.tsx
@@ -46,8 +46,8 @@ export function withAuthenticator(
     includeGreetings = false,
     authenticatorComponents = [],
     federated = null,
-    theme = null
-    ) {
+    theme = null,
+    signUpConfig = {}) {
     return class extends Component<IAuthenticatorProps, IWithAuthenticatorState> {
         constructor(props) {
             super(props);
@@ -58,6 +58,20 @@ export function withAuthenticator(
                 authState: props.authState || null,
                 authData: props.authData || null
             };
+
+            this.authConfig = {};
+
+            if (typeof includeGreetings === 'object' && includeGreetings !== null){
+                this.authConfig = Object.assign(this.authConfig, includeGreetings);
+            } else {
+                this.authConfig = {
+                    includeGreetings,
+                    authenticatorComponents,
+                    federated,
+                    theme,
+                    signUpConfig
+                };
+            }
         }
 
         handleAuthStateChange(state, data) {
@@ -71,11 +85,11 @@ export function withAuthenticator(
                 return (
                     <div>
                         {
-                            includeGreetings?
+                            this.authConfig.includeGreetings?
                             <Greetings
                                 authState={authState}
                                 authData={authData}
-                                federated={federated || this.props.federated}
+                                federated={this.authConfig.federated || this.props.federated || {} }
                                 onStateChange={this.handleAuthStateChange}
                                 theme={theme}
                             />
@@ -93,11 +107,12 @@ export function withAuthenticator(
 
             return <Authenticator
                 {...this.props}
-                theme={theme}
-                federated={federated || this.props.federated}
-                hideDefault={authenticatorComponents.length > 0}
+                theme={this.authConfig.theme}
+                federated={this.authConfig.federated || this.props.federated}
+                hideDefault={this.authConfig.authenticatorComponents && this.authConfig.authenticatorComponents.length > 0}
+                signUpConfig={this.authConfig.signUpConfig}
                 onStateChange={this.handleAuthStateChange}
-                children={authenticatorComponents}
+                children={this.authConfig.authenticatorComponents || []}
             />;
         }
     };