+++ packages/aws-amplify-vue/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.2.1-unstable.0"></a>
+## [0.2.1-unstable.0](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.1.11-unstable.0...aws-amplify-vue@0.2.1-unstable.0) (2018-12-10)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-vue
+
 <a name="0.1.11-unstable.0"></a>
 ## [0.1.11-unstable.0](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.1.10...aws-amplify-vue@0.1.11-unstable.0) (2018-12-07)
 
@@ -72,12 +80,10 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
-
 **Note:** Version bump only for package aws-amplify-vue
 
 <a name="0.1.8-unstable.3"></a>
 ## [0.1.8-unstable.3](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.1.8-unstable.2...aws-amplify-vue@0.1.8-unstable.3) (2018-11-30)
-
 
 
 
@@ -88,12 +94,10 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
-
 **Note:** Version bump only for package aws-amplify-vue
 
 <a name="0.1.8-unstable.1"></a>
 ## [0.1.8-unstable.1](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.1.8-unstable.0...aws-amplify-vue@0.1.8-unstable.1) (2018-11-21)
-
 
 
 
@@ -104,12 +108,29 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+**Note:** Version bump only for package aws-amplify-vue
+
+<a name="0.2.1-beta.3"></a>
+## [0.2.1-beta.3](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.2.1-beta.2...aws-amplify-vue@0.2.1-beta.3) (2018-11-19)
+
+
+**Note:** Version bump only for package aws-amplify-vue
+
+<a name="0.2.1-beta.2"></a>
+## [0.2.1-beta.2](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.2.1-beta.1...aws-amplify-vue@0.2.1-beta.2) (2018-11-14)
+
+
+**Note:** Version bump only for package aws-amplify-vue
+
+<a name="0.2.1-beta.1"></a>
+## [0.2.1-beta.1](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.1.7-unstable.1...aws-amplify-vue@0.2.1-beta.1) (2018-11-14)
+
+
 
 **Note:** Version bump only for package aws-amplify-vue
 
 <a name="0.1.7"></a>
 ## [0.1.7](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.1.7-unstable.1...aws-amplify-vue@0.1.7) (2018-11-12)
-
 
 
 
@@ -128,13 +149,17 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+**Note:** Version bump only for package aws-amplify-vue
+
+<a name="0.2.1-beta.0"></a>
+## [0.2.1-beta.0](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.1.6-unstable.0...aws-amplify-vue@0.2.1-beta.0) (2018-11-02)
+
+
 
 **Note:** Version bump only for package aws-amplify-vue
 
 <a name="0.1.6"></a>
 ## [0.1.6](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.1.6-unstable.0...aws-amplify-vue@0.1.6) (2018-11-01)
-
-
 
 
 **Note:** Version bump only for package aws-amplify-vue